pyproject.toml
[tool.poetry]
name = "excel-files-parser"
version = "0.1.0"
description = "Excel files parser"
authors = ["Zeev <zeev.soskin@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.41"
python-dotenv = "^1.1.1"
pydantic = "^2.11.7"
psycopg2-binary = "^2.9.10"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
